backend/pyproject.toml
[project]
name = "lexigrok"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Tim Dumol", email = "tim@timdumol.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi[all]>=0.116.1",
    "minio>=7.2.15",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "pydantic-settings>=2.0.0",
    "openai>=1.97.0",
    "elevenlabs>=2.7.1",
    "sqlmodel>=0.0.24",
]

[project.scripts]
lexigrok = "lexigrok:main"

[dependency-groups]
dev = [
    "ruff",
    "pre-commit",
    "pytest>=8.4.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["uv_build>=0.7.20,<0.8.0"]
build-backend = "uv_build"
